pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reviewer2"
version = "1.0.1"
description = "Adversarial peer review pipeline for academic PDFs."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "The Catalogue of Errors Ltd"}]
keywords = ["peer-review", "llm", "academic", "pdf"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
]
dependencies = [
    "pypdf",
    "requests",
    "tiktoken>=0.7.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "pydantic>=2.0",
    "python-dotenv (>=1.2.2,<2.0.0)",
]

[project.optional-dependencies]
gemini = [
    "google-genai",
]
# docling (PDF-to-markdown) is not listed here because its torch transitive
# dependency makes automated resolution fragile. Install it separately:
#   pip install docling
# The pipeline degrades gracefully if docling is absent.
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
reviewer2 = "reviewer2.cli:main"

[project.urls]
Homepage = "https://github.com/CasparDP/reviewer2"
Issues = "https://github.com/CasparDP/reviewer2/issues"
Source = "https://github.com/CasparDP/reviewer2"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
reviewer2 = [
    "prompts/*.txt",
    "prompts/resources/*.txt",
    "prompts/resources/*.pdf",
    "data/*.csv",
]
